(#1406) Move `FallbackFrom` to `org.cactoos.func`
