+++ src/main/java/org/cactoos/func/FallbackFrom.java
@@ -21,11 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.scalar;
+package org.cactoos.func;
 
 import org.cactoos.Func;
 import org.cactoos.iterable.IterableOf;
 import org.cactoos.iterable.Mapped;
+import org.cactoos.scalar.InheritanceLevel;
+import org.cactoos.scalar.MinOf;
 
 /**
  * Fallback from exception.
