+++ src/main/java/org/cactoos/func/FuncWithFallback.java
@@ -27,7 +27,6 @@ import java.sql.SQLException;
 import java.sql.SQLRecoverableException;
 import org.cactoos.Func;
 import org.cactoos.iterable.IterableOf;
-import org.cactoos.scalar.FallbackFrom;
 import org.cactoos.scalar.InheritanceLevel;
 import org.cactoos.scalar.ScalarWithFallback;
