+++ src/main/java/org/cactoos/iterable/IterableOf.java
@@ -28,10 +28,10 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import org.cactoos.Func;
 import org.cactoos.Scalar;
+import org.cactoos.func.FallbackFrom;
 import org.cactoos.func.UncheckedFunc;
 import org.cactoos.iterator.IteratorOf;
 import org.cactoos.scalar.And;
-import org.cactoos.scalar.FallbackFrom;
 import org.cactoos.scalar.Folded;
 import org.cactoos.scalar.Or;
 import org.cactoos.scalar.ScalarWithFallback;
