+++ src/test/java/org/cactoos/func/FallbackFromTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.scalar;
+package org.cactoos.func;
 
 import java.io.IOException;
 import java.util.IllegalFormatException;
