+++ src/test/java/org/cactoos/func/FuncWithFallbackTest.java
@@ -27,7 +27,6 @@ import java.io.IOException;
 import java.util.IllegalFormatException;
 import java.util.IllegalFormatWidthException;
 import org.cactoos.iterable.IterableOf;
-import org.cactoos.scalar.FallbackFrom;
 import org.junit.Test;
 import org.llorllale.cactoos.matchers.Assertion;
 import org.llorllale.cactoos.matchers.FuncApplies;
